Working hybrid, caveat wss server is created in same file that middleware is defined, have not yet successfullly shared wss server object across files (mvc
